third_party/blink/web_tests/external/wpt/editing/run/backcolor-expected.txt

This is a testharness.js-based test.
Found 42 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span></p><p> <span style=\\"background-color:rgb(0, 255, 255)\\"><span>bar</span></span> </p><p><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span></p>" but got "<p><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span></p><p> <span style=\\"background-color:rgb(0, 255, 255)\\"><span>bar</span> </span></p><p><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span></p>"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span></p><p> <span style=\\"background-color:rgb(0, 255, 255)\\"><span>bar</span></span> </p><p><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span></p>" but got "<p><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span></p><p> <span style=\\"background-color:rgb(0, 255, 255)\\"><span>bar</span> </span></p><p><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("backcolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 255, 255)" but got "rgba(0, 0, 0, 0)"
[FAIL] [["backcolor","#00FFFF"]] "<p style=\\"background-color: rgb(0, 255, 255)\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foobarbaz</p>" but got "<p style=\\"\\"><span style=\\"background-color:rgb(0, 255, 255)\\">foobarbaz</span></p>"
[FAIL] [["backcolor","#00FFFF"]] "<p style=\\"background-color: #00ffff\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foobarbaz</p>" but got "<p style=\\"\\"><span style=\\"background-color:rgb(0, 255, 255)\\">foobarbaz</span></p>"
[FAIL] [["backcolor","#00FFFF"]] "<p style=\\"background-color: aqua\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foobarbaz</p>" but got "<p style=\\"\\"><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span><span style=\\"background-color:rgb(0, 255, 255)\\">bar</span><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "{<p style=\\"background-color: aqua\\">foo</p><p>bar</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foo</p><p><span style=\\"background-color:rgb(0, 255, 255)\\">bar</span></p>" but got "<p style=\\"\\"><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span></p><p><span style=\\"background-color:rgb(0, 255, 255)\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "{<p style=\\"background-color: aqua\\">foo</p><p>bar</p>}" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "{<p style=\\"background-color: aqua\\">foo</p><p>bar</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foo</p><p><span style=\\"background-color:rgb(0, 255, 255)\\">bar</span></p>" but got "<p style=\\"\\"><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span></p><p><span style=\\"background-color:rgb(0, 255, 255)\\">bar</span></p>"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "{<p style=\\"background-color: aqua\\">foo</p><p>bar</p>}" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<span style=\\"background-color: aqua\\">foo<span style=\\"background-color: tan\\">[bar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(0, 255, 255)\\">foobarbaz</span>" but got "<span style=\\"background-color:rgb(0, 255, 255)\\">foo</span><span style=\\"background-color:rgb(0, 255, 255)\\">bar</span><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span>"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<span style=\\"background-color: aqua\\">foo<span style=\\"background-color: tan\\">[bar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(0, 255, 255)\\">foobarbaz</span>" but got "<span style=\\"background-color:rgb(0, 255, 255)\\">foo</span><span style=\\"background-color:rgb(0, 255, 255)\\">bar</span><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span>"
[FAIL] [["backcolor","#00FFFF"]] "<span style=\\"background-color: aqua\\">foo<span style=\\"background-color: tan\\">b[ar]</span>baz</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(0, 255, 255)\\">foo<span style=\\"background-color:rgb(210, 180, 140)\\">b</span>arbaz</span>" but got "<span style=\\"background-color:rgb(0, 255, 255)\\">foo</span><span style=\\"background-color:rgb(210, 180, 140)\\">b</span><span style=\\"background-color:rgb(0, 255, 255)\\">ar</span><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span>"
[FAIL] [["backcolor","#00FFFF"]] "<p style=\\"background-color: aqua\\">foo<span style=\\"background-color: tan\\">b[ar]</span>baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"background-color:rgb(0, 255, 255)\\">foo<span style=\\"background-color:rgb(210, 180, 140)\\">b</span>arbaz</p>" but got "<p style=\\"\\"><span style=\\"background-color:rgb(0, 255, 255)\\">foo</span><span style=\\"background-color:rgb(210, 180, 140)\\">b</span><span style=\\"background-color:rgb(0, 255, 255)\\">ar</span><span style=\\"background-color:rgb(0, 255, 255)\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<div style=\\"background-color: aqua\\"><p style=\\"background-color: tan\\">b[ar]</p></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"background-color:rgb(0, 255, 255)\\"><p style=\\"background-color:rgb(210, 180, 140)\\">b<span style=\\"background-color:rgb(0, 255, 255)\\">ar</span></p></div>" but got "<div style=\\"\\"><p style=\\"\\"><span style=\\"background-color:rgb(210, 180, 140)\\">b</span><span style=\\"background-color:rgb(0, 255, 255)\\">ar</span></p></div>"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<div style=\\"background-color: aqua\\"><p style=\\"background-color: tan\\">b[ar]</p></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"background-color:rgb(0, 255, 255)\\"><p style=\\"background-color:rgb(210, 180, 140)\\">b<span style=\\"background-color:rgb(0, 255, 255)\\">ar</span></p></div>" but got "<div style=\\"\\"><p style=\\"\\"><span style=\\"background-color:rgb(210, 180, 140)\\">b</span><span style=\\"background-color:rgb(0, 255, 255)\\">ar</span></p></div>"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<span style=\\"display: block; background-color: aqua\\"><span style=\\"display: block; background-color: tan\\">b[ar]</span></span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"display:block; background-color:rgb(0, 255, 255)\\"><span style=\\"display:block; background-color:rgb(210, 180, 140)\\">b<span style=\\"background-color:rgb(0, 255, 255)\\">ar</span></span></span>" but got "<span style=\\"display:block\\"><span style=\\"display:block\\"><span style=\\"background-color:rgb(210, 180, 140)\\">b</span><span style=\\"background-color:rgb(0, 255, 255)\\">ar</span></span></span>"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<span style=\\"display: block; background-color: aqua\\"><span style=\\"display: block; background-color: tan\\">b[ar]</span></span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"display:block; background-color:rgb(0, 255, 255)\\"><span style=\\"display:block; background-color:rgb(210, 180, 140)\\">b<span style=\\"background-color:rgb(0, 255, 255)\\">ar</span></span></span>" but got "<span style=\\"display:block\\"><span style=\\"display:block\\"><span style=\\"background-color:rgb(210, 180, 140)\\">b</span><span style=\\"background-color:rgb(0, 255, 255)\\">ar</span></span></span>"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "fo[o<span style=background-color:tan>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<span style=\\"background-color:rgb(0, 255, 255)\\">o</span><span style=\\"background-color:rgb(210, 180, 140)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span>baz" but got "fo<span style=\\"background-color:rgb(0, 255, 255)\\">ob</span><span style=\\"background-color:rgb(210, 180, 140)\\">ar</span>baz"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "fo[o<span style=background-color:tan>b]ar</span>baz" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "fo[o<span style=background-color:tan>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<span style=\\"background-color:rgb(0, 255, 255)\\">o</span><span style=\\"background-color:rgb(210, 180, 140)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span>baz" but got "fo<span style=\\"background-color:rgb(0, 255, 255)\\">ob</span><span style=\\"background-color:rgb(210, 180, 140)\\">ar</span>baz"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "fo[o<span style=background-color:tan>b]ar</span>baz" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "foo<span style=background-color:tan>ba[r</span>b]az" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"background-color:rgb(210, 180, 140)\\">ba<span style=\\"background-color:rgb(0, 255, 255)\\">r</span></span><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>az" but got "foo<span style=\\"background-color:rgb(210, 180, 140)\\">ba</span><span style=\\"background-color:rgb(0, 255, 255)\\">rb</span>az"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "foo<span style=background-color:tan>ba[r</span>b]az" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "foo<span style=background-color:tan>ba[r</span>b]az" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"background-color:rgb(210, 180, 140)\\">ba<span style=\\"background-color:rgb(0, 255, 255)\\">r</span></span><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>az" but got "foo<span style=\\"background-color:rgb(210, 180, 140)\\">ba</span><span style=\\"background-color:rgb(0, 255, 255)\\">rb</span>az"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "foo<span style=background-color:tan>ba[r</span>b]az" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["backcolor","#00FFFF"]] "fo[o<span style=background-color:tan>bar</span>b]az" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "foo[<span style=background-color:tan>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"background-color:rgb(210, 180, 140)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span>baz" but got "foo<span style=\\"background-color:rgb(0, 255, 255)\\">b</span><span style=\\"background-color:rgb(210, 180, 140)\\">ar</span>baz"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "foo[<span style=background-color:tan>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"background-color:rgb(210, 180, 140)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span>baz" but got "foo<span style=\\"background-color:rgb(0, 255, 255)\\">b</span><span style=\\"background-color:rgb(210, 180, 140)\\">ar</span>baz"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "foo<span style=background-color:tan>ba[r</span>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"background-color:rgb(210, 180, 140)\\">ba<span style=\\"background-color:rgb(0, 255, 255)\\">r</span></span>baz" but got "foo<span style=\\"background-color:rgb(210, 180, 140)\\">ba</span><span style=\\"background-color:rgb(0, 255, 255)\\">r</span>baz"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "foo<span style=background-color:tan>ba[r</span>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"background-color:rgb(210, 180, 140)\\">ba<span style=\\"background-color:rgb(0, 255, 255)\\">r</span></span>baz" but got "foo<span style=\\"background-color:rgb(210, 180, 140)\\">ba</span><span style=\\"background-color:rgb(0, 255, 255)\\">r</span>baz"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<span style=background-color:tan>fo[o</span><span style=background-color:yellow>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(210, 180, 140)\\">fo<span style=\\"background-color:rgb(0, 255, 255)\\">o</span></span><span style=\\"background-color:rgb(255, 255, 0)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span>" but got "<span style=\\"background-color:rgb(210, 180, 140)\\">fo</span><span style=\\"background-color:rgb(0, 255, 255)\\">ob</span><span style=\\"background-color:rgb(255, 255, 0)\\">ar</span>"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<span style=background-color:tan>fo[o</span><span style=background-color:yellow>b]ar</span>" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<span style=background-color:tan>fo[o</span><span style=background-color:yellow>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(210, 180, 140)\\">fo<span style=\\"background-color:rgb(0, 255, 255)\\">o</span></span><span style=\\"background-color:rgb(255, 255, 0)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span>" but got "<span style=\\"background-color:rgb(210, 180, 140)\\">fo</span><span style=\\"background-color:rgb(0, 255, 255)\\">ob</span><span style=\\"background-color:rgb(255, 255, 0)\\">ar</span>"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<span style=background-color:tan>fo[o</span><span style=background-color:yellow>b]ar</span>" queryCommandIndeterm("backcolor") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<span style=background-color:tan>fo[o</span><span style=background-color:tan>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(210, 180, 140)\\">fo<span style=\\"background-color:rgb(0, 255, 255)\\">o</span></span><span style=\\"background-color:rgb(210, 180, 140)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span>" but got "<span style=\\"background-color:rgb(210, 180, 140)\\">fo</span><span style=\\"background-color:rgb(0, 255, 255)\\">ob</span><span style=\\"background-color:rgb(210, 180, 140)\\">ar</span>"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<span style=background-color:tan>fo[o</span><span style=background-color:tan>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(210, 180, 140)\\">fo<span style=\\"background-color:rgb(0, 255, 255)\\">o</span></span><span style=\\"background-color:rgb(210, 180, 140)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span>" but got "<span style=\\"background-color:rgb(210, 180, 140)\\">fo</span><span style=\\"background-color:rgb(0, 255, 255)\\">ob</span><span style=\\"background-color:rgb(210, 180, 140)\\">ar</span>"
[FAIL] [["stylewithcss","true"],["backcolor","#00FFFF"]] "<span style=background-color:tan>fo[o<span style=background-color:transparent>b]ar</span></span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(210, 180, 140)\\">fo<span style=\\"background-color:rgb(0, 255, 255)\\">o</span><span style=\\"background-color:rgba(0, 0, 0, 0)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span></span>" but got "<span style=\\"background-color:rgb(210, 180, 140)\\">fo</span><span style=\\"background-color:rgb(0, 255, 255)\\">ob</span><span style=\\"background-color:rgba(0, 0, 0, 0)\\">ar</span>"
[FAIL] [["stylewithcss","false"],["backcolor","#00FFFF"]] "<span style=background-color:tan>fo[o<span style=background-color:transparent>b]ar</span></span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"background-color:rgb(210, 180, 140)\\">fo<span style=\\"background-color:rgb(0, 255, 255)\\">o</span><span style=\\"background-color:rgba(0, 0, 0, 0)\\"><span style=\\"background-color:rgb(0, 255, 255)\\">b</span>ar</span></span>" but got "<span style=\\"background-color:rgb(210, 180, 140)\\">fo</span><span style=\\"background-color:rgb(0, 255, 255)\\">ob</span><span style=\\"background-color:rgba(0, 0, 0, 0)\\">ar</span>"
Harness: the test ran to completion.
